// lims-frontend/src/types/apps/globalReferenceTypes.ts
export type GlobalReferenceType = {
    createdAt: string;
    updatedAt: string | null;
    createdBy: string;
    updatedBy: string | null;
    id: number;
    kode_kategori: string | null;
    kode: string | null;
    deskripsi1: string | null;
    deskripsi2: string | null;
    grouping: number | null;
    kode_kategori_display: string | null;
    uuid: string;
}